requirements/base.txt
Flask==2.3.3
gspread==5.11.2
google-auth==2.23.3
google-auth-oauthlib==1.2.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
